requirements.txt
discord.py
chatterbot
chatterbot_corpus
nest_asyncio
transformers
torch
groq
requests
